+++ libde265/encoder/encpicbuf.cc
@@ -134,7 +134,7 @@ void image_data::set_references(int sps_index, // -1 -> custom
   //shdr.num_ref_idx_l1_active = l1.size();
 
   assert(l0.size() < MAX_NUM_REF_PICS);
-  for (int i=0;i<l0.size();i++) {
+  for (size_t i=0;i<l0.size();i++) {
     shdr.RefPicList[0][i] = l0[i];
   }
 
@@ -245,7 +245,7 @@ void encoder_picture_buffer::mark_encoding_finished(int frame_number)
 
 bool encoder_picture_buffer::have_more_frames_to_encode() const
 {
-  for (int i=0;i<mImages.size();i++) {
+  for (size_t i=0;i<mImages.size();i++) {
     if (mImages[i]->state < image_data::state_encoding) {
       return true;
     }
@@ -257,7 +257,7 @@ bool encoder_picture_buffer::have_more_frames_to_encode() const
 
 image_data* encoder_picture_buffer::get_next_picture_to_encode()
 {
-  for (int i=0;i<mImages.size();i++) {
+  for (size_t i=0;i<mImages.size();i++) {
     if (mImages[i]->state < image_data::state_encoding) {
       return mImages[i];
     }
@@ -269,7 +269,7 @@ image_data* encoder_picture_buffer::get_next_picture_to_encode()
 
 const image_data* encoder_picture_buffer::get_picture(int frame_number) const
 {
-  for (int i=0;i<mImages.size();i++) {
+  for (size_t i=0;i<mImages.size();i++) {
     if (mImages[i]->frame_number == frame_number)
       return mImages[i];
   }
@@ -281,7 +281,7 @@ const image_data* encoder_picture_buffer::get_picture(int frame_number) const
 
 image_data* encoder_picture_buffer::get_picture(int frame_number)
 {
-  for (int i=0;i<mImages.size();i++) {
+  for (size_t i=0;i<mImages.size();i++) {
     if (mImages[i]->frame_number == frame_number)
       return mImages[i];
   }
@@ -293,7 +293,7 @@ image_data* encoder_picture_buffer::get_picture(int frame_number)
 
 bool encoder_picture_buffer::has_picture(int frame_number) const
 {
-  for (int i=0;i<mImages.size();i++) {
+  for (size_t i=0;i<mImages.size();i++) {
     if (mImages[i]->frame_number == frame_number)
       return true;
   }
